adfsdfdfg

+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: lvasseur <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/07 13:19:10 by lvasseur          #+#    #+#             */
-/*   Updated: 2016/11/11 15:53:07 by lvasseur         ###   ########.fr       */
+/*   Updated: 2016/11/11 15:55:30 by lvasseur         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while ((unsigned char)s1[i] == (unsigned char)s2[i] && s1[i] && s2[i]
